requirements.txt
google-generativeai>=0.8.0
tavily-python>=0.5.0
rich>=13.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
